[core] Split Interpolator into a separate file
